% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_as_job.R
\name{render_job}
\alias{render_job}
\title{render an rmarkdown file in background using RStudio Jobs}
\usage{
render_job(input, params = list(), output_file)
}
\arguments{
\item{input}{.rmd document to be knitted}

\item{params}{params to pass to the .rmd}

\item{output_file}{name of the output_file (and the job)}
}
\description{
if you want to
}
\examples{
\dontrun{
   render_job("document.Rmd", list(dataset = "df1"), "summary_df1.html")
}

}
