% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity.R
\name{view_in_excel}
\alias{view_in_excel}
\title{Open in Excel}
\usage{
view_in_excel(x)
}
\arguments{
\item{x}{a dataframe to open in Excel}
}
\description{
Simple helper, so I don't complain about the slugginess of RStudio's View so much
}
\examples{
\dontrun{
view_in_excel(Titanic)
}
}
